test: Tweaked test rig a little

+++ tests/e2e_all/scripts/common/stop_daemons.sh
@@ -9,15 +9,17 @@ source "$testScriptsDir/common/check_env.include.sh" || exit $?
 
 for pid in $(pgrep -f "sshnpd.*${commitId}")
 do
+  echo "Killing $(ps -ef | grep " $pid " | grep -v grep)"
   kill "$pid"
 done
 
-for pid in $(pgrep -f "tests/e2e_all/runtime/.*/.* ")
+for pid in $(pgrep -f "tests/e2e_all/(runtime|releases)/.*/.* ")
 do
-  kill "$pid"
-done
-
-for pid in $(pgrep -f "tests/e2e_all/releases/.*/.* ")
-do
-  kill "$pid"
+  processInfo=$(ps -ef | grep " $pid " | grep -v grep)
+  # Allow the srv's to exit in their own time
+  if ! grep -q "srv " <<< "$processInfo";
+  then
+    echo "Killing $processInfo"
+    kill "$pid"
+  fi
 done
